Twiggle Builders
* Squashed commit of the following:

commit 9fb7833abb182a066803484774de706b586ad250
Merge: c641512 41f1f9d
Author: Owenvii <owen9456@gmail.com>
Date:   Thu Jul 7 16:04:36 2022 -0600

    Merge branch 'main' into GeneralRefactoring

commit c6415128bb03afea43b1c379693b8d0f2612f279
Author: Owenvii <owen9456@gmail.com>
Date:   Thu Jul 7 16:01:35 2022 -0600

    Introduced a constant for packet ID's.

    * Removed hard coding for packet ID's by moving their strings to a static final variable
    * Moved some subclasses in the ItemBuyResponse packet to their own classes in respective packages.

commit 317a979d8e7465c3a3e11886b9face9c8f6740c0
Author: Owenvii <owen9456@gmail.com>
Date:   Wed Jul 6 17:46:24 2022 -0600

    Trying to implement 'loading in' logic for player online status?

commit 2a3c7d0d789cf057e3c576075d3cf1fe577dd8ae
Author: Owenvii <owen9456@gmail.com>
Date:   Wed Jul 6 16:43:53 2022 -0600

    Update initial furniture loading

    Update initial furniture loading to use the new OI packet style.

commit 2698923af36f4c4f18351f613893efacdbfe60cb
Author: Sky Swimmer <stefaneekhof@aerialworks.ddns.net>
Date:   Wed Jul 6 18:16:49 2022 +0200

    Fix that the kick event is fired even if the player is online

commit fe574bb3f19fca9342c70a9ab1ddc53d44bcdd0a
Author: Owenvii <owen9456@gmail.com>
Date:   Tue Jul 5 13:10:16 2022 -0600

    Added a object delete packet. Updated player to use the object delete packet.

commit d6ccb938546e566b8e28b1695085f69c6ccb8ece
Author: Owenvii <owen9456@gmail.com>
Date:   Tue Jul 5 12:59:10 2022 -0600

    update Player.syncTo to use a new OI packet made for players

* Fix Build

* Re-Implement Cleanups

Oops, merging from master really does hit different

* Improve cross-world player warp

* Implement the chat filter

* Added a chat command for raw packet sending

* Filter update

* Joining and leaving minigames

* Twiggle Builders

+++ src/main/java/org/asf/emuferal/packets/xt/gameserver/sanctuaries/SanctuaryJoinPacket.java
@@ -21,6 +21,7 @@ public class SanctuaryJoinPacket implements IXtPacket<SanctuaryJoinPacket> {
 	
 	public String sanctuaryOwner = null;
 	public int mode = 0;
+	private boolean isAllowed;
 
 	@Override
 	public String id() {
@@ -72,45 +73,41 @@ public class SanctuaryJoinPacket implements IXtPacket<SanctuaryJoinPacket> {
 			// Verify access
 			if (privSetting == 2) {
 				// Nobody
+				isAllowed = false;
 
-				// Send error
-				client.sendPacket("%xt%rj%-1%false%1689%2%-1%" + sanctuaryOwner + "%sanctuary_" + sanctuaryOwner + "%");
-				return true;
 			} else if (privSetting == 1) {
 				// Followers
-
 				// Check if the owner follows the current player
 				if (!SocialManager.getInstance().getPlayerIsFollowing(sanctuaryOwner, player.account.getAccountID())) {
-					// Send error
-					client.sendPacket(
-							"%xt%rj%-1%false%1689%2%-1%" + sanctuaryOwner + "%sanctuary_" + sanctuaryOwner + "%");
-					return true;
+					isAllowed = false;
 				}
 			}
 		}
 
 		// Build room join
 		JoinRoom join = new JoinRoom();
+		join.success = isAllowed;
 		join.levelType = 2;
 		join.levelID = 1689;
 		join.roomIdentifier = "sanctuary_" + sanctuaryOwner;
 		join.teleport = sanctuaryOwner;
 
-		// Sync
-		GameServer srv = (GameServer) client.getServer();
-		for (Player plr2 : srv.getPlayers()) {
-			if (plr2.room != null && player.room != null && player.room != null && plr2.room.equals(player.room)
-					&& plr2 != player) {
-				player.destroyAt(plr2);
+		if (isAllowed == true){
+			// Sync
+			GameServer srv = (GameServer) client.getServer();
+			for (Player plr2 : srv.getPlayers()) {
+				if (plr2.room != null && player.room != null && player.room != null && plr2.room.equals(player.room)
+						&& plr2 != player) {
+					player.destroyAt(plr2);
+				}
 			}
+
+			// Assign room
+			player.roomReady = false;
+			player.pendingLevelID = 1689;
+			player.pendingRoom = "sanctuary_" + sanctuaryOwner;
+			player.levelType = join.levelType;
 		}
-
-		// Assign room
-		player.roomReady = false;
-		player.pendingLevelID = 1689;
-		player.pendingRoom = "sanctuary_" + sanctuaryOwner;
-		player.levelType = join.levelType;
-
 		// Send packet
 		client.sendPacket(join);
